new explore command (old predict)
