+++ src/biome/text/commands/__init__.py
@@ -1,4 +1,3 @@
-from .explore import BiomeExplore
 from .learn import BiomeLearn
-from .predict import BiomePredict
+from .explore import BiomeExplore
 from .serve import BiomeRestAPI
