--- src/biome/text/commands/predict/__init__.py
@@ -1,1 +1,0 @@
-from .predict import BiomePredict
